Add missing CaloConst includes, minox fixes

+++ DbTables/inc/CalLaserTimeCalib.hh
@@ -8,6 +8,7 @@
 #include <sstream>
 #include <map>
 #include "Offline/DbTables/inc/DbTable.hh"
+#include "Offline/DataProducts/inc/CaloConst.hh"
 #include "Offline/DataProducts/inc/CaloSiPMId.hh"
 
 namespace mu2e {
@@ -17,7 +18,7 @@ namespace mu2e {
 
       class Row {
         public:
-        Row(CaloSiPMId  roid, double T0, double ErrT0, double chisq, int nev):
+        Row(CaloSiPMId  roid, float T0, float ErrT0, float chisq, int nev):
           _roid(roid),_T0(T0),_ErrT0(ErrT0),_chisq(chisq),_nev(nev) {}
         CaloSiPMId       roid()     const { return _roid;} // Offline ID
         float     T0()     const { return _T0; }
@@ -74,7 +75,6 @@ namespace mu2e {
 
     private:
       std::vector<Row> _rows;
-      //std::map<int,std::size_t> _chanIndex;
   };
 
 }
